controllers

+++ src/controllers/videoController.js
@@ -1,4 +1,7 @@
 import Video from "../models/Video";
+import Comment from "../models/Comment";
+import User from "../models/User";
+import { async } from "regenerator-runtime";
 
 /*
 Video.find({}, (error, videos) => {
@@ -8,14 +11,16 @@ Video.find({}, (error, videos) => {
 
 export const home = async (req, res) => {
   // Video.find({}) -> {} 안이 비어있으면 모든 형식을 찾는다는 것을 뜻한다.
-  const videos = await Video.find({}).sort({ createdAt: "desc" });
+  const videos = await Video.find({})
+    .sort({ createdAt: "desc" })
+    .populate("owner");
   return res.render("home", { pageTitle: "Home", videos });
 };
 
 export const watch = async (req, res) => {
   const { id } = req.params;
   //=> const id = req.params.id;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner").populate("comments");
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
   }
@@ -24,26 +29,42 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
+  }
+  // video의 owner의 id가 로그인 한 user의 id와 같은지 확인하는 작업
+  if (String(video.owner) !== String(_id)) {
+    req.flash("error", "Not authorized");
+    return res.status(403).redirect("/");
   }
   return res.render("edit", { pageTitle: `Edit: ${video.title}`, video });
 };
 
 export const postEdit = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
   const { id } = req.params;
   const { title, description, hashtags } = req.body;
-  // Model.exists(_id : id) --> id의 존쟈 여부만을 확인한다.-> 아랫줄에 !video가 true인지 false인지 나타내기 위해서
+  // Model.exists(_id : id) --> id의 존재 여부만을 확인한다.-> 아랫줄에 !video가 true인지 false인지 나타내기 위해서
   const video = await Video.exists({ _id: id });
   if (!video) {
+    req.flash("error", "You are not the owner of the video");
     return res.render("404", { pageTitle: "Video not found" });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   await Video.findByIdAndUpdate(id, {
     title,
     description,
     hashtags: Video.formatHashtags(hashtags),
   });
+  req.flash("success", "Changes saved.");
   return res.redirect(`/videos/${id}`);
 };
 
@@ -52,13 +73,23 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
+  const { video, thumb } = req.files;
   const { title, description, hashtags } = req.body;
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
+      fileUrl: video[0].path,
+      thumbUrl: thumb[0].path,
+      owner: _id,
       hashtags: Video.formatHashtags(hashtags),
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/");
   } catch (error) {
     return res.status(400).render("upload", {
@@ -70,6 +101,16 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await User.findById(id);
+  if (!video) {
+    return res.render("404", { pageTitle: "Video not found" });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   // Video.findByIdAndRemove()도 있지만 특별한 이유가 없는 이상 findByIdAndDelete를 사용()
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
@@ -84,7 +125,59 @@ export const search = async (req, res) => {
         // i는 ignore의 약자로 대소문자 무시를 의미한다
         $regex: new RegExp(keyword, "i"),
       },
-    });
+    }).populate("owner");
   }
   return res.render("search", { pageTitle: "Search", videos });
 };
+
+export const registerView = async (req, res) => {
+  const { id } = req.params;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.sendStatus(404);
+  }
+  video.meta.views = video.meta.views + 1;
+  await video.save();
+  return res.sendStatus(200);
+};
+
+export const createComment = async (req, res) => {
+  const {
+    session: { user },
+    body: { text },
+    params: { id },
+  } = req;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.sendStatus(404);
+  }
+  const comment = await Comment.create({
+    text,
+    owner: user._id,
+    video: id,
+  });
+  video.comments.push(comment._id);
+  video.save();
+  return res.status(201).json({ newCommentId: comment._id });
+};
+
+// export const removeComment = async (req, res) => {
+//   const {
+//     session: {
+//       user: { _id },
+//     },
+//     params: { id },
+//   } = req;
+
+//   const comment = await Comment.findById(id).populate("owner");
+
+//   const video = await Video.findById(id);
+//   if (!video) {
+//     return res.sendStatus(404);
+//   }
+//   if (String(comment.owner._id) !== String(_id)) {
+//     return res.status(403).redirect("/");
+//   }
+//   await Comment.findByIdAndDelete(id);
+//   return res.status(201).json({ newCommentId: id });
+// };
